bot/pyproject.toml
[project]
name = "lms-bot"
version = "0.1.0"
description = "Telegram bot for LMS"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
